requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Cython==0.28.5
Django==2.1
django-extensions==2.1.0
djangorestframework==3.8.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
pyliblo==0.10.0
pytz==2018.5
requests==2.19.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.23